Plug an infinite loop when grabbing blocks.
The problem is that a piece might be 'in progress', but all blocks
said piece has is accounted for and being downloaded by different
clients. In that case there will be no blocks to grab and the splitAt
call won't work.

The result is we continually grab the same piece again and again, yielding
an infinite loop in the process.