Remove the OMBox structure.
Rationale: The OMBox structure were originally needed because of
connect-backs to the Peer manager from individual peers and that I
thought it had to be the Peer manager which choked and unchoked clients.
The way it worked, there was an obvious deadlock bug lurking in the
dark.

With the choking/unchoking done by the PieceMgr however, there is little
need for this structure. It also solves some pesky Leak problems at the
same time.