Add supervision support to the peer processes.
This set of changes updates the Peer Process code and its containing
processes to use a supervisor process to maintain themselves. This is
part of an ongoing effort to shuffle control of the process hierarchy
towards the supervisor model.

We also work towards to linking together the Choke manager and Peer
processes properly. This will still take some changes to completely get
to work.

Conflicts:
	src/PeerP.hs