Enable Status messaging from the Peer process.
This update enables that a peer will contact the status process every 30
seconds and then update it with how many bytes were uploaded and
downloaded in the client. This is then propagated via the status
process towards the tracker process when the tracker is contacted the
next time.

We can now report the right amounts to the tracker.