Track the version of Haskell-torrent.
Impose a configure script on the build process. This script tries to use
git to figure out what version we are currenctly on and then it
generates a description if possible. Otherwise, we just take a default
value.

Also, mark the commit id as dirty if somebody changed the source code in
any way before building. These changes greatly simplifies the
bug-tracking process.