Implement endgame support.
When we enter endgame, we use the list of downloading pieces to grab
pieces from. This list is then sent to the Peer and the peer must handle
the situation accordingly.

Add support for cancelling Requests in Peers. Introduce endgame
broadcasts in PieceMgr. Introduce endgame support in the Choke Mgr.

Add Piece Manager assertions. At each round, we currently check the
state of the piece database for several invariants to check that it does
not corrupt.