Audit ChokeMgr.
This change introduces several changes to the ChokeMgr process. It
mainly cleans up and pushes the Transformer stack further into the call
stack. This in turn enables better logging of what is going on in the
process and is the first step towards improving the state.

While here, add some simple assertions on how upload slots are behaving
when the system is running correctly.