Correct choking considerations.
When we pick peers eligible for the choking and unchoking, we do not
want:

  - Peers which are seeders. Unchoking them gives us nothing, as they
    already have all of the file.
  - Among the leechers, we do not want peers not interested in us. Since
    they are not interested, there is nothing to be gained by unchoking
    them as we don't expect to be able to transfer data to them.

In practice, we unchoke all of the second kind. This is to ensure that
if they get interested, they get the oppurtunity right away. But when
choosing peers to unchoke, they are not relavant.

This patch fixes a wrong boolean operator such that the above happens.