When putting back blocks, ignore completed.
A problem arises when we want to put back blocks while other Peers are
running in endgame mode. The other peers might end up completing a piece
before the non-endgaming process gives up and puts back its blocks. Then
the putback code would see that there is no piece in the progress map it
could use for adding the blocks to.

The first rule is, that if the piece is already completely downloaded,
we don't care. Second, if it is still there, in progress, we can add the
putback-blocks on top of it.