Fix a bug in the cereal WireProtocol code.
There was a bug in the WireProtocol code based on cereal. The bug was
that we did not encode the size of the packet as a 32-bit big endian
number in front of the packet in question. The result was that other
clients silently closed the socket on us.

With this change, we can merge Alex Masons cereal work to the master
branch soonish.