Create a Protocol hierarchy.
Several people encouraged to use a Hier-module space rather than a flat
space currently used. We will, in the course of several commits, move
modules around and place them in the hier where we belong they should
be.

It may be that these positions are wrong and have to be corrected in the
long run.