Make the Piece knowledge of a peer an IntSet.
Again, profiling shows this speeds up the code by a fair amount. It is
the block grabber which is the CPU-hog at the moment. It can be
optimized, but it requires another approach than what we have right now.
In particular, it requires a better handling of the Piece database.