Implement support for handling incoming conns.
When a connection is incoming, we handle it by looking if we need more
connections or not. If we do, we try to initiate a hanshake with the
peer in the hope it will provide us with another peer. If the handshake
succeeds, we add the peer as usual business to the Peer managers group
of peers.