Propagate Seeder knowledge to the Choke Manager.
This change improves the system as a Seeder tremendously. Beforehand, we
always regarded a peer to be a leecher and thus always eligible for
unchoking. With this change, we skip every seeder when choosing whom to
unchoke. The seeders should just send us data. It is the leechers we
want to serve with our precious bandwidth.

Reflect this in the TODO file.