Fix a channel connection bug. Plug a memory leak.
Since we coalesced the ChokeInfo Channel and the ChokeChannel into a
single one, the PieceMgr was queueing up messages on a channel which
never got any event to fire. This led to a memory leak when we had a
large number of messages piling up on that channel. This change fixes
this problem and I have verified that the memory usage stays low.