Improve performance by killing an PieceSet.intersection.
The code first produced a PieceSet with the sole purpose of running an
Intersection on it. This makes sense in a world where we try to use Sets
for all operations, but in practice it turns out to be a rather slow
plan.

The main culprit for a profiling run was the tryGrabProgress code which
will be called whenever a peer wants some more blocks to download. The
change is to use supplied pieceset is a fast check (It is O(1)) and then
bail out as soon as something hits.

Preliminary runs suggest this improves the codes performance.