Don't recalculate the size of the pieceset in peers.
Rather than recalculate the size of the pieceset whenever a peer sends
us a have message (potentially some 8000 bitshift operations and a lot
of memory usage in the process), we just keep an external counter of
what is missing and update that counter whenever the peer sends us a new
message.

isASeeder then becomes a counter lookup.