Use Arrays for PieceMaps rather than Data.Map.
Since we don't dynamically update the PieceMap when it has been built,
there is little reason to keep it in a map and do rather expensive
lookups in the map that may fail. Furthermore, we don't have to keep
Data.Map Tree nodes around anymore, which should improve memory usage
quite a bit.