Represent Digests as ByteStrings
There are two positive outcomes: ByteStrings are faster for most
manipulation. Furthermore, we have 20-element Strings. These are 16
bytes, so that means 20*16 = 320 bytes per piece. Suppose we have 4096
pieces. Then that is 1310720 bytes, i.e., 1280 Kilobytes. With the
change, it is a mere 80K.

It is a substantial change with many torrents loaded of large size.