Improve the splitter into seeders and leechers.
The rules governing are now:

If the peer is a seeder, we are not interested in unchoking him. He
already has everything.

If the peer is not interested in us, we won't be considering him here.
All non-interested gets unchoked automatically so they are ready if they
become interested.

Now, if the peer survives these checks, he may be on a torrent we seed.
In that case he is put into the seeders box. Otherwise, we go on.

If the peer is snubbing us, we don't want him in the hot-seat.

Otherwise, we add him to the leecher set.