Split the timerTick into several pieces. Optimize.
Splitting and profiling shows we allocate a *lot* of data in
timerTickChokeMgr. The only place where this allocation can sensibly
take place is in the atomically block, or in the rate extraction code.

The atomically block now doesn't allocate underneath its transaction,
which ought to mean contention will not hurt.