Return interest on Have messages to the PieceMgr.
When we tell the PieceMgr that a peer HAVE certain pieces, we check
interest on them right away. We then return the list of interesting
pieces directly to the peer, so the peer can work with them.

In this first step, we simply throw away the information. But in later
steps we plan to use this information to remove the considerInterest
calls in peers.