Minimize the size of the histogram.
We only need to track pieces in the histogram which are interesting.
Pieces which are not interesting are already downloaded and thus we can
just skip them in the histogram.

Also, whenever we complete a piece, we kill it in the histogram. The
histogram thus get smaller and smaller over time and it will be really
small while a torrent is initially propagating.